+++ libs/components/src/lib/components/dialogs/confirm-dialog/confirm-dialog.module.ts
@@ -5,12 +5,14 @@ import { PrizmOverlayModule } from '../../../modules/overlay';
 import { PrizmButtonModule } from '../../button';
 import { PrizmDialogConfirmComponent } from './confirm-dialog.component';
 import { PrizmScrollbarModule } from '../../scrollbar';
+import { PrizmThemeModule } from '@prizm-ui/theme';
 
 @NgModule({
   imports: [
     CommonModule,
     PolymorphModule,
     PrizmOverlayModule,
+    PrizmThemeModule,
     PrizmButtonModule,
     PrizmFocusTrapModule,
     PrizmScrollbarModule,