+++ libs/components/src/lib/components/dialogs/sidebar/sidebar.module.ts
@@ -6,11 +6,13 @@ import { PrizmButtonModule } from '../../button';
 import { PrizmSidebarComponent } from './sidebar.component';
 import { PrizmScrollbarModule } from '../../scrollbar';
 import { PrizmInputIconButtonModule } from '../../input/common/input-icon-button/input-icon-button.module';
+import { PrizmThemeModule } from '@prizm-ui/theme';
 
 @NgModule({
   imports: [
     CommonModule,
     PolymorphModule,
+    PrizmThemeModule,
     PrizmOverlayModule,
     PrizmInputIconButtonModule,
     PrizmButtonModule,